dashboard-admin/backend/analytics/requirements.txt
fastapi
uvicorn[standard]
clickhouse-driver
asyncpg
python-dotenv
PyJWT